Added data and data analysis tool screenshots to new directory
track_creation_data to be used for track creation (used UV18, last time we used the asymmetric oval with good documentation, and is a good example of a simplified FSAE track)

Other data examples to show how data will be exported, especially AEM data that we will be transitioning to. However, templates can be made for those csv exports as desired.